Setup trace hooks for libmctp
This commit setups up Hostboot to hook into the pr_* trace commands
that the libmctp core logic uses. This is helpful for debugging
in some cases. Notice that this does not leverage the tracehash
features the rest of our traces use so we must be judicious in
how many extra traces we add to the libmctp code.

Change-Id: I975641303117d21d9e94b49f240acefe70190be4
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/116339
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Reviewed-by: Matthew Raybuck <matthew.raybuck@ibm.com>
Reviewed-by: Zachary Clark <zach@ibm.com>
Reviewed-by: Isaac Salem <isaac.salem@ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>