Trustedboot: Fix Special Case of Extending and Logging the HBB Section
The HBBL code extends the Hostboot Base Code (HBB) PNOR Section into
the primary TPM's PCR0.  When Hostboot code takes over the IPL it needs
to log this extension to PCR0 while also both extending and logging this
data into the primary TPM's PCR1.

This commit fixes a bug in how this special case was being handled.  The
code was erronesouly *not* extending the HBB code to PCR1 but was logging it.

Change-Id: Ifcaa3e2d46fd61a4c682df52636b8a209ec3214a
CQ:SW523591
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/116362
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Reviewed-by: Deborah L McLemore <debmc@us.ibm.com>
Reviewed-by: Roland Veloz <rveloz@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>
Reviewed-by: Christopher J Engel <cjengel@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>