Move Internode Comm Area
Internode comm area was on the bottom of the HB reserved memory.
In the case when memory mirroring is enabled, all mirrored memory
ranges (starting with arch reg and ending with the TCE table) are
doubled, which overruns and erases the internode comm area that was
on the bottom of the res mem range. Erasing the node comm area
made boot hang in MPIPL with mirrored memory enabled, since secondary
nodes could not read the correct state from the erased memory. This
commit moves the Internode Comm Area higher right before arch reg data
(where the mirrored memory starts).

Change-Id: Idea45ed64bad40dd61be64a4803bf0dcc20976af
CQ: SW524080
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/116595
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>
Reviewed-by: Zachary Clark <zach@ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Chen Du <duchen@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>