Speed up "hb prime" process
The first run of the "hb prime" command has to clone git repositories,
build tools, etc. On subsequent invocations much of this work can be
skipped, and some can be parallelized. This commit makes several
changes to speed up priming the hostboot sandbox, particularly on
the second invocation and after:

- Parallelize PNOR partition signing
- Don't rebuild FFS tools on every "prime", and use multiple processes
  parallel when we do build them
- Fix SBE build redundancy detection (previously the code checked for
  existence of files that were never being created)
- Add a new flag "HB_FAST_PRIME" that activates all "prime" fast paths
  (currently only implies HB_SKIP_GEN_HCODE)

These changes, used with HB_SKIP_GEN_HCODE/HB_FAST_PRIME, bring the
time needed to "prime" a standalone sandbox down from about 15 minutes
to about 2 minutes on average.

The new official way to invoke the optimized path in "hb prime", which
can be used when only Hostboot changes are being made (as opposed to
HCODE or other changes), is:

  HB_FAST_PRIME=1 vexec hb prime -j32

Change-Id: Ie80b2cd30694deeba1534ababc779a60e359c849
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/116667
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Matthew Raybuck <matthew.raybuck@ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>