Support LPC over XSCOM for alternate boot processor
DD1.0 and DD2.0 chips have an LPC MMIO hardware bug that triggers
checkstop on LPC errors.  This firmware workaround implements an LPC
driver that uses the XSCOM LPC facility to bypass the hardware error
at the expense of performance.  This XSCOM LPC facility will be used
on the alternate boot processor only.  On any LPC error with this
facility, the recovery will reset the entire LPC complex.

Change-Id: Idd1bbf5a9d14f12c1a4a542e3d9ca0bf04e18b07
CQ: SW523870
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/116797
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Christian R Geddes <crgeddes@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>