Support for I2C Lock Contention with the SBE
With the introduction of PMIC telemetry gathering, there is the
possibility for contention on the i2c bus out to the DDIMM.
The two users are the SBE doing the PMIC operations and HBRT
writing to the SPD for a memory repair.  To avoid problems, we
will utilize the Atomic Lock feature of the I2C Master on P10.
Because HBRT uses PHYP for the i2c operations, the lock handling
itself will be done within PHYP.  On the HBRT side we need to
handle any error scenarios that are specific to our usage.

A second change needs to be made to force clear the lock as part
of our SBE reset behavior.  This ensures that the logic is in an
unlocked state while the SBE is dead/resetting.

Change-Id: I5214fce8d73485066e79459cbca6eaafdb42b783
RTC: 281971
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/116954
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Matthew Raybuck <matthew.raybuck@ibm.com>
Reviewed-by: Corey V Swenson <cswenson@us.ibm.com>
Reviewed-by: Michael Baiocchi <mbaiocch@us.ibm.com>