Fix for defect SW518715
The TOD sync_spread() routine can fail due to an excessive delay
between the start of the SCOM writes to the TOD Timer registers
and the end of these SCOM writes.  This excessive delay is
caused by the inherent process timesharing nature of the FSP
Linux OS.  This fix moves the configuration of the TOD Timer
registers to before the TOD FSMs enter the Running State,
instead of after.

Change-Id: I86bfdff41717459870b04b4ea4550a4d1fcaf5a4
CQ: SW518715
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/116843
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: HWSV CI <hwsv-ci+hostboot@us.ibm.com>
Reviewed-by: Douglas Holtsinger <douglas.holtsinger@ibm.com>
Reviewed-by: Joseph J McGill <jmcgill@us.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/116847
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>