Call shutdownDuringIPL in handleShutdownMsg if iv_syncPointReached is set
If Hostboot received a shutdown request from the FSP while the node its
running on is waiting at the sync point, and the shutdown request
originated from another node TI'ing, Hostboot will not honor the shutdown
request. This is because a feature was added to ensure we reach the same
Istep where the original TI'ing node failed. This is a problem when
Hostboot is waiting at the sync point because the check to see if
Hostboot has reached the appropriate Istep is done in the doIstep
function which only runs while Hostboot is driving the IPL. At the sync
point the FSP is driving the IPL so we will never hit this check to
trigger the shutdown.

Change-Id: I5b242fdbf1c735937e9312195a53ba80e7a3bf4e
CQ: SW525693
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/116960
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Christian R Geddes <crgeddes@us.ibm.com>
Reviewed-by: Isaac Salem <isaac.salem@ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>