Assume TPM not present if TDP bit is set.
Previously if the TDP bit for the primary TPM was set we would
set hwasState.present to match the FSP's presence determination.
This is a problem because if the TDP bit being set causes a deconfig
when TPMs are required forcing a reconfig at the end of
host_discover_targets. This results in a fail over without gard of
the failing TPM. If the backup TPM is also failing we will end up
in an endless reboot because the TPM never gets garded. To get around
this we will assume the TPM is not present if the TDP bit is set
which, if the TPM is detected by the FSP, will cause a mismatch and
the TPM will get garded.

CQ: SW526441
Change-Id: Iebda5be3fdf3d56cf0b3f496c141eec805e611ee
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/117543
Reviewed-by: Michael Baiocchi <mbaiocch@us.ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>