PLDM Watchdog: Pause Isteps
This commit introduces logic to be able to pause isteps for a specified
amount of time or indefinitely, depending on the settings of a special
attribute. While the isteps are paused and in PLDM mode, the code will
periodically send out a heartbeat to BMC to indicate that the host is
still up. This is done in lieu of us not being able to disable PLDM
watchdog. The period of heartbeats during the wait is the min watchdog
timeout / 2.

Change-Id: Ib4e02f7636f90b93e66345218f011b1848e297d2
RTC: 250776
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/117622
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Reviewed-by: Zachary Clark <zach@ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>