Evaluate presence after checkCRC in platPresenceDetect
In the platPresenceDetect function there are extra CRC checks
done on SPD the OCMB targets, if there is a mismatch we want to
mark the OCMB as not present. This code already exists but it was
not removing the target from the list of present targets that
this function returns. This commit moves the logic that removes
not present targets from the list to the end of the target loop
to ensure the OCMB is marked not present for CRC mismatches.

Change-Id: I9934deea0fe1f82dfd14481dabb4ff7807a57bfe
CQ: SW526418
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/117681
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Chen Du <duchen@us.ibm.com>
Reviewed-by: Isaac Salem <isaac.salem@ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>