Fix position number for fapi attribute dump tool used by cronus
Add three new attributes to all targets, FAPINAME_NODE, FAPINAME_POS,
and FAPINAME_UNIT which will be populated in processMrw.pl with the
correct values corresponding to the numbers in being used currently
when we set the FAPI_NAME attribute. Use these new attributes when
a fapi attribute dump is requested as this is what the cronus tools
are expecting.

Change-Id: If918e439b84487d6312baace9a08bae37dfdcba6
CQ: SW525769
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/117756
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Matthew Raybuck <matthew.raybuck@ibm.com>
Reviewed-by: Chen Du <duchen@us.ibm.com>
Reviewed-by: Isaac Salem <isaac.salem@ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>