Force cache refresh for module vpd contents
Hostboot keeps a cache of all of vpd and other eeprom content
that we consume during the ipl.  This vastly improves boot time
since the hardware accesses can be quite slow.  The data is
only refreshed if we notice a serial or part number change.

However, if another entity modifies the VPD we can end up with
a stale cache, or even worse we can corrupt the data due to
overlapping ECC.  To address this we will now be refreshing
the module vpd contents on every boot.

Change-Id: Ib99c759f55f4c1860b9e28ab179da1a2b5e2abb1
CQ: SW521541
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/117859
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Zachary Clark <zach@ibm.com>
Reviewed-by: Isaac Salem <isaac.salem@ibm.com>
Reviewed-by: Christian R Geddes <crgeddes@us.ibm.com>