Read data from hardware for partial page writes
The SPI controller in P10 does all seeprom writes in
256 byte pages. When we want to only update a portion
of that data we have to read in the remainder in order
to write out a complete page.

Previously we would read the extra data from our
cache, but that has shown problems due to other
entities (FSP) updating other records without Hostboot
knowing.  This change will switch to reading the
extra data directly from the hardware instead to
allow us to work around changes to unrelated portions
of the module vpd.

Change-Id: Ic986a15e940cf81d20c3481a3462bd8a46bc24aa
CQ: SW521541
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/117930
Reviewed-by: Zachary Clark <zach@ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Isaac Salem <isaac.salem@ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Christian R Geddes <crgeddes@us.ibm.com>