Re-enable redundant TOD topology and enforce fabric restrictions
Everest has a maximum of 8 processors in a single system. Because of
limited IOHS hardware, each processor is connected to the other
processors by "half" or "split" IOHS links, i.e. by one SMPGROUP
connection.

Because of this split link setup, the TOD topology setup was
previously failing on Everest systems when initializing the secondary
topology, and this was halting the IPL, so we temporarily disabled
secondary topology initialization to work around the problem.

Research determined that the problem is that, due to the quirks of the
split SMP topology, only the first processor on a given DCM can be
used as a TOD master. Previously the code was using processor 1 on DCM
0 as the master for the secondary topology, which was causing the
failure.

This commit re-enables secondary topology setup and addresses the
original problem by blacklisting all secondary processors for the TOD
master selection process.

Change-Id: I59ca639bb9a8211a5bf51a5bab5def28296aa4e0
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/117959
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>
Tested-by: Daniel M Crowell <dcrowell@us.ibm.com>