Update OCC sensor ID encoding
This commit updates the OCC sensor ID encoding to support the new PLDM
and DBUS relationships. The OCC sensor ID will now indicate the type
of target (CORE or DIMM) and which PLDM entity ID corresponds to the
target. This will allow the BMC to figure out the DBUS path from the
sensor ID.

Change-Id: I6007f35b407ecc5a4a5dad3f3413eb687300b3da
RTC: 282003
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/118085
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Reviewed-by: Christopher J Cain <cjcain@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>