PLDM: Add support to consume hb_lib_ids PLDM bios attribute.
This commit adds the code to read and parse the hb_lid_ids attribute
and calls that function during the Pnor Resouse Provider initialization.
This attribute tells us which lid id maps to the legacy pnor section
eyecatchers hostboot is familiar with. The Pnor Resource Provider will
store this information and use it to translate between lid ids and
legacy pnor section ids.

Change-Id: Ibb708c719cc5b3b05ff8c36ae0bbedb7c45eec50
RTC: 208802
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/118087
Reviewed-by: Isaac Salem <isaac.salem@ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>