Add new hwas plat function to call ECO mode HWP during discover targets
Adds a new platform function to call the HWP responsible for determining
which cores to use for Extended Cache-Only mode. Since this is common
code a platform function is necessary to allow HWSV to do the same in a
way consistent with their platform.

This call is done in discover targets since we want to call that HWP as
soon as we can and before we start enforcing specific rules on CORE
targets. The restrictEcUnits function is responsible for some of these
rule enforcements and thus the location of the plat function call was
placed just prior to the first time that function is called to ensure
ATTR_ECO_MODE has been setup on the functional COREs.

This commit doesn't implement handling of ECO cores, just begins the
groundwork to allow for that handling in a future commit.

Change-Id: I9d698122fb910e9d9586c720b6cb9c7ddd8528e0
RTC:282723
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/118088
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Reviewed-by: Roland Veloz <rveloz@us.ibm.com>
Reviewed-by: Christian R Geddes <crgeddes@us.ibm.com>
Reviewed-by: Michael Baiocchi <mbaiocch@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>