Add support for Extended Cache-Only (ECO) mode
There exists a distinction between regular cores and cores to be used for
ECO mode. This commit adds hostboot support for that distinction and a
few functions that will filter based on ATTR_ECO_MODE to utilFilter.

The support for ECO cores mostly boils down to ignoring or working
around some rules that are in place for regular cores as they don't
apply to ECO cores.

Change-Id: I9b2e8157ae1323282a753fdfee37499b6eaff53e
RTC:282723
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/118160
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Reviewed-by: Michael Baiocchi <mbaiocch@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>