Create PLDM shutdown callback manager and refactor FRU state updates
This commit defines an API in the PLDM module that allows registering
callbacks that need to be invoked when the system begins to shut down.

This commit also moves PLDM FRU functional states synchronization from
the Initservice code into one of these PLDM shutdown callbacks, which
is the preferred way to handle shutdown tasks because it keeps the
Initservice code clean and allows easy dependency management between
shutdown tasks.

Change-Id: I484298368990ca75fdcfedf6eeaa7b9c5baaf5c9
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/118245
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Reviewed-by: Christian R Geddes <crgeddes@us.ibm.com>