Power Line Disturbance support for BMC during Hostboot IPL
This commit ensure Hostboot waits at least 10 seconds after an errorlog
was initially created before we process the callouts and send the
log to the BMC. When an error log is created its constructor reads
the timebase register and stores it. When the error log is being
processed by the errlmanager we will check the current time and wait
until at least 10 seconds have transpired since the log was created.
This will allow the BMC to detect and handle any PLD problems that
might occur. We only want this delay for logs with callouts which
might lead to an unnecessary service request. This commit also
introduces a new system attribute called DISABLE_PLD_WAIT which can
be set to 1 via attribute overrides to disable this 10 second delay.
HBRT will require a separate solution which will come in another commit.

Change-Id: I9ed31e591de8355ac95002cd90793c0b67dc3c30
RTC: 267588
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/118545
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>
Reviewed-by: Matthew Raybuck <matthew.raybuck@ibm.com>
Reviewed-by: Isaac Salem <isaac.salem@ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>