Add state query handler for shutdown sensor
Previously, Hostboot distinguished between state sensors that
represented a queryable state, and state sensors that were only used
in PLDM notifications to the BMC. The former kind had state query
handlers so that the BMC could request their state, but handlers were
not needed or implemented for the latter kind.

Now, the eBMC has started requesting the states of all state sensors
at certain points in the IPL. To accomodate this, this commit adds
query handlers for all kinds of state sensors (in particular the
shutdown state sensor) and removes the API support to be able to
register an unqueryable state sensor at all.

This commit also removes the error path that is triggered when the BMC
requests the state of an OCC sensor too early in the IPL. Now,
Hostboot will respond with a "sensor state unknown" status on the OCC
sensors.

Change-Id: Ic6d0d8c3f2d96debb6f4df763f61c5f3580feaeb
CQ: SW527723
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/118577
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Reviewed-by: Christian R Geddes <crgeddes@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>