Power Line Disturbance support for BMC during Hostboot Runtime
This commit ensures Hostboot runtime waits at least 10 seconds after
an error is initially created before we process any callouts and
send the log onto the BMC. This is done by having the errlmanager at
runtime do a nanosleep for 10 seconds if the log has any callouts
that might lead to a maintenance request. To disable this wait the
DISABLE_PLD_WAIT attribute can be overriden to be set to 1.

Change-Id: Ibfce49f7245d3ad143831e8153e4ee449957ab2b
RTC: 267588
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/118666
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>
Reviewed-by: Isaac Salem <isaac.salem@ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>