Fix for defect SW521078
Return a new error code for TOD secondary topology failures to distinguish
it from all other failures, but only if the secondary topology failure
is the only failure.  Any other failure will cause an assertion
first before the end of the HWP.  Continue to the end of the HWP
if the secondary topology fails and no other failure is seen.
A secondary topology failure can still cause a checkstop,
if configured to do so, as the TOD error register is not masked.

Change-Id: I4cee7dfe7e5d24e611dfa743a70080695f7f0dfe
Original-Change-Id: If86f370c4774719ce8073491077a5253e719adf4
CQ:SW521078
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/114044
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Thi N Tran <thi@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: HWSV CI <hwsv-ci+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Joachim Fenkes <fenkes@de.ibm.com>
Reviewed-by: Zachary Clark <zach@ibm.com>
Reviewed-by: DAVID A KING <kingboy@us.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/129571
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>
Tested-by: Daniel M Crowell <dcrowell@us.ibm.com>