Improve the previous bug fix in genPnorImages.pl image signing parallelization
The previous bug fix added a mechanism in genPnorImages.pl to serialize
its calls to crtSignedContainer.sh when in production mode because
crtSignedContainer.sh can't be run in parallel due to filename collisions.
This commit also serializes the calls in the case of development-to-production
drivers as this case also requires calling crtSignedContainer.sh in
production mode.

Change-Id: I22379a0081afd89f4eced0d515ac407d5657d4d6
CQ:SW527882
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/118673
Reviewed-by: Zachary Clark <zach@ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>