Update HB_NFS_DIR checker script and fail more gracefully
Simics has updated their mini-python binary to use python 3. As a part
of python 3 print statements now require parentheses around the args.

This commit updates validate-hb-nfs-dir.py to reflect that requirement
and also updates the denali simics scripts for autobooting to fail more
gracefully when validate-hb-nfs-dir.py fails. Since that script is not a
requirement for auto-patching and auto-booting of denali simics.

Change-Id: If53f67f609ddd335f20f19f3bcd87f28459ab11a
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/118678
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Reviewed-by: Isaac Salem <isaac.salem@ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Chen Du <duchen@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>