Pending BIOS Attributes: Implement SetBIOSAttributeCurrentValue API (commit 1/n)
FUNCTIONAL OVERVIEW

There are certain PLDM BIOS attributes that control system IPL
behavior whose values are only sampled and used once at the beginning
of the IPL. Though these attributes only have an effect at the
beginning of the IPL when they are sampled by host firmware, their
values can be changed at any time using PLDM commands (and these
settings will only take effect on the next IPL). Therefore, when a
component in the system needs to know what the currently active value
of one of these attributes is (i.e. what a value was when host
firmware sampled it at the start of the current IPL), it cannot use
the actual current value of the attributes, because they could have
changed since firmware IPLed and sampled it.

To provide access to the original value of these "latched" attributes
after the system IPLs, each "logical" latched attribute X is split
into two "actual" PLDM BIOS attributes named X_current and X. The
attribute X_current contains what the value of the attribute was when
the system IPLed (i.e. the value which is actually in effect for the
current IPL). The attribute X (with no suffix) contains the value that
will be sampled on the next IPL.

When a system component wants to alter the logical latched attribute
X, it will write to the PLDM BIOS attribute X of the same name (NOT
X_current). Then at the beginning of each IPL, host firmware will
sample the attribute X and store its value in the attributed
X_current. It will inform the BMC of this by sending a
SetBIOSAttributeCurrentValue request for the X_current attribute with
the new value from attribute X. When the system needs to know the
currently effective value of the logical latched attribute X, it will
read the value of the PLDM BIOS attribute X_current. In other words,
the host firmware reads X and writes X_current, whereas everything
else reads X_current and writes X.

THIS COMMIT

To support this scheme, this commit adds an API to perform the
SetBIOSAttributeCurrentValue PLDM request. This will be used to
implement attribute value latching for pending+current BIOS attribute
pairs in a future commit.

Change-Id: I4710a9c393f3bfd1829896066b2e68de3c57410a
RTC: 284192
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/118681
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Reviewed-by: Christian R Geddes <crgeddes@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>
Tested-by: Daniel M Crowell <dcrowell@us.ibm.com>