Dont deconfig PROCs that dont have WOF tables
In the case that WOF data is bad there are two options:
1. Deconfigure the PROC with bad data and let the rest run at WOF speeds
2. Disable WOF on the whole system but use all available PROCs

This commit changes the code to take the second option because if the
first option is taken then the system will lose more than just the PROC.
It will lose all the memory, IO, and--for multinode--the entire drawer.

Going forward, Hostboot expects the PM code to never fail out due to an
error back from WOF accessors. Instead, the log will just be committed
and downgrade things as needed to allow the system to IPL.

Additionally, this means that if a system has only PROCs without WOF
data on them then they can still IPL.

Change-Id: I03b5f29c65e59a56d0794a83fe59a5fa9a1ce324
CQ:SW528182
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/118691
Reviewed-by: Chen Du <duchen@us.ibm.com>
Reviewed-by: Isaac Salem <isaac.salem@ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>