Check AttrOverrideSync Message Before Processing
A crash can occur during HB shutdown path where we could
get an attribute sync message that can be deleted before
we get a chance to process it. This causes us to not complete
the shutdown correctly. This change adds a check for non-nullptr-
ness of the incoming message before processing it.

Change-Id: I691fb93293719e88c59265d6cd39974fd6441f8a
CQ: SW526098
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/118688
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Reviewed-by: Corey V Swenson <cswenson@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>