Fix for SRCs using devdesc as their custdesc
Before this change, if an error log was missing a customer description,
then genErrlParsers.pl would append the developer description to some
default text and use that as the custdesc. This is bad since the devdesc
could contain sensitive IBM internal info. For this change, genErrlParsers.pl
will now use "An internal firmware error occurred." for error logs as the
default custdesc text if one is not provided.

Change-Id: I73bbca2b84613359ab5710457182dbe2956aeac0
CQ: SW528194
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/118726
Reviewed-by: MARISSA GARZA <garzam@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Zane C Shelley <zshelle@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>